Move booksform to containers, remove categories from initialstate and exchange the options in booksform for categories for a .map function instead
